requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
langchain_community
pyPDF2
chromadb